ccMain v1.1
An update to the ccMain file. Includes the first upgrade, a clicker, an 'unlocked structure' functionality to hide structures until their predecessor has been bought a certain amount of times (i.e. you cannot see the 'goblin tribe' structure until you have purchased 3 'cookie goblins'), and a new structure, the aforementioned 'goblin tribe'. The code has also been refactored to improve readability and its ability to be modified in the future; information is now loaded from the user info file in the beginning of the program with an if conditional instead of through the Cookie class.